feat(test): one npm test, four URL-driven categories
