+++ .github/workflows/_test.yml
@@ -8,9 +8,6 @@ on:
       ignore-scripts:
         type: boolean
         default: false
-      protocol:
-        # RPC, WS
-        type: string
       node:
         # Juno, Pathfinder
         type: string
@@ -41,7 +38,6 @@ jobs:
 
     env:
       USE_DEVNET: ${{ inputs.use-devnet }}
-      IS_WS: ${{ inputs.protocol == 'WS' }}
       NODE: ${{ inputs.node == 'Juno' && 'juno' || inputs.node == 'Pathfinder' && 'pathfinder' || '' }}
       VERSION: ${{ inputs.version }}
 
@@ -65,7 +61,8 @@ jobs:
         # are built from the Starknet Foundation base domain:
         #   https://sepolia.<domain>/<node>/rpc/<version>
         #   wss://sepolia.<domain>/<node>/ws/<version>
-        # The WS suites also drive an account, so an rpc url is always needed.
+        # Both are always emitted: the runner decides what to run from the urls alone, and an
+        # rpc url is needed in every case — the WebSocket suites drive an account over HTTP.
         run: |
           if [ "$USE_DEVNET" = 'true' ]; then
             rpc_url='http://127.0.0.1:5050/rpc'
@@ -73,11 +70,9 @@ jobs:
           elif [ -n "$SNF_NODE_URL" ] && [ -n "$NODE" ] && [ -n "$VERSION" ]; then
             base="sepolia.$SNF_NODE_URL/$NODE"
             rpc_url="https://$base/rpc/$VERSION"
+            ws_url="wss://$base/ws/$VERSION"
             echo "::add-mask::$rpc_url"
-            if [ "$IS_WS" = 'true' ]; then
-              ws_url="wss://$base/ws/$VERSION"
-              echo "::add-mask::$ws_url"
-            fi
+            echo "::add-mask::$ws_url"
           else
             echo "::error::No node url available: set use-devnet, or provide the node and version inputs together with the SNF_NODE_URL secret."
             exit 1
@@ -86,15 +81,11 @@ jobs:
           echo "ws-url=$ws_url" >> "$GITHUB_OUTPUT"
       - id: run-tests
         name: run tests
-        run: |
-          args=()
-          # The WS run is restricted to the WebSocket suites; the `--` lets npm forward the
-          # path filter through to Jest. It used to also need --forceExit, because dropping a
-          # still-subscribed socket left a connection the gateway would not release and Jest
-          # never exited; the suite now unsubscribes that socket itself, so a run that hangs
-          # here is a real leak and must not be papered over.
-          [[ "${{ inputs.protocol }}" == "WS" ]] && args+=( '__tests__/WebSocket' )
-          npm run test:coverage -- "${args[@]}"
+        # No path filter and no protocol flag: a single run covers every category the resolved
+        # urls allow — unit once, the node-touching suites once per transport, and the WebSocket
+        # suites on top. `--forceExit` must not come back either; the suites close their own
+        # sockets, so a run that hangs here is a real leak and must not be papered over.
+        run: npm run test:coverage
         env:
           TEST_RPC_URL: ${{ steps.node-url.outputs.rpc-url }}
           TEST_WS_URL: ${{ steps.node-url.outputs.ws-url }}
