+++ __tests__/config/index.ts
@@ -5,6 +5,7 @@ export * from './helpers/env';
 export * from './helpers/initDevnetHistory';
 export * from './helpers/logging';
 export * from './helpers/waitUntil';
+export * from './helpers/wsTestHelpers';
 export * from './helpers/mockTransport';
 export * from './helpers/mockWebSocket';
 export * from './schema';
