+++ __tests__/config/projects/http.ts
@@ -1,0 +1,2 @@
+// The default transport, stated explicitly so a run's category is visible from its config alone.
+process.env.TEST_TRANSPORT = 'http';
